A bunch of work to get the test suite running smoothly.
Some of the tests would cause StackOverflows or other crashes that hung up the test runner. The test runner would also choke when writing out the nunit results if there where control characters from the re & unicode tests. Finally, running multithreaded causes the runner to deadock, so the thread count for the TC runs is set to 1.

There are some fixes in the tests, but mostly this is just to get it running and checking results.