Add a Build.proj file as the top-level build script.
IronPython lacks a top-level build script, which makes building more complex
than it needs to be. The file will build all of the available targets. It
also includes a task to build a Stage directory that packages can be built
from.

Finally, there is now a Platform directory that holds the assemblies for each supported platform.