Fix #31420, #31736: Centralize the version information in the build system.
The version info will be stored in the build system and the appropriate
constants generated at build time so that the version only has to be changed
in one place.

The only file that needs to be edited to change the IronPython version is Build.proj. This will generate CurrentVersion.Generated.cs for IronPython.dll, and the other IronPython assemblies plus the PythonContext get their version information from it. Only private methods were changed, so the public API is unbroken.

The DLR assemblies are unaffected by these changes.

This also adds a set of custom tasks in
Solutions/CustomTasks/CustomTasks.cs, which is built by Bootstrap.proj. All
of the projects have been modified to ensure that this is built before
anything else. The code generation for the version information depends on
these custom tasks.