Complete the rewrite of the IronPython installer.
The new installer has the same features as the old one, but is cleaned up so
that it is easier to maintain. Newer WiX features like automatic IDs remove
a lot of the noise from the wix files.

GAC installation is now optional. In fact, most stuff is now optional, but
installed by default.

The IronPython Tools for Visual Studio have been removed. Upgrading to
IronPython 2.7.2 will remove them. Use Python Tools for Visual Studio
instead.