Implement BZ2Decompressor.
This required some rather hackish changes to BZip2InputStream to be able to
save and restore the current state, so that it could be rewound after and
error and then retried when new data came in. It's ugly, but it works.

Performance likely suffers as a result of this, but I haven't measured it
yet. Not doing a save/restore for every byte would likely speed things up a
bit.

It also ends up buffering the entire input in memory; overwriting the
input buffer after successful decoding should avoid that, and it's possible
since the BZip2InputStream never looks back.