Fix installation when IronRuby is installed.
When IronRuby is installed, the Microsoft.* assemblies in the GAC are
not overwritten on installation of IronPython because the AssemblyVersion
is the same. Despite this, IronPython requires the newer version with the
same version number :|. The easiest way to fix this is to bump the
AssemblyFileVersion, which will tell the MSI engine to replace a GAC'd
assembly even though the AssemblyVersion hasn't changed.